[users] fix: Checed Journalists group before update for saving user

+++ back/apps/users/models.py
@@ -50,8 +50,9 @@ def pre_save_user(sender, instance: User,  **kwargs):
 @receiver(post_save, sender=User)
 def post_save_user(sender, instance: User, **kwargs):
     g = Group.objects.get(name='Journalists')
-    if instance.type in User.TYPE.EDITORS:
-        if not instance.groups.filter(id__in=[g.id]).exists():
-            g.user_set.add(instance)
-    else:
-        instance.groups.filter(id__in=[g.id]).delete()
+    if g:
+        if instance.type in User.TYPE.EDITORS:
+            if not instance.groups.filter(id__in=[g.id]).exists():
+                g.user_set.add(instance)
+        else:
+            instance.groups.filter(id__in=[g.id]).delete()
